Songeditor: fix keyboard hide issuee
